Check completeness of lines before interpreting
For cases like chaining and else statements that
can be on a newline but must be associated with a
previous statement, interpreting line by line will not work.

To get around this, at every line x, the next line will
be parsed to check completeness and if it needs to be
appeneded to x to be valid.