Add in basic displayers
Add displayer support for jvm objects, to start this just covers
DataSet/DataFrames. Calling display(obj) on any object from the kernel
will return a full bundle with as rich a display as possible. The default
from the Displayers class is to create a rich plaintext. The customized
addition for Dataset/dataframe includes html and data resource mimetypes
as well.